Fixed a warning

+++ src/core/modules/filters/filters_recipients.h
@@ -81,7 +81,7 @@ class IRecipientFilterExt
 public:
 	static int __getitem__(IRecipientFilter* pFilter, unsigned index)
 	{
-		if (index >= pFilter->GetRecipientCount())
+		if (index >= (unsigned int) pFilter->GetRecipientCount())
 			BOOST_RAISE_EXCEPTION(PyExc_IndexError, "Index out of range.")
 
 		return pFilter->GetRecipientIndex(index);
